Bug 898216 - Change the credential lookup strategy when the USE_DEV_KEY_LOOKUP environment variable is set to true
