+++ env.sample
@@ -41,3 +41,8 @@ export USE_LAZY_ADMIN=false
 
 # Login server URL without auth for webmaker SSO
 export LOGIN_SERVER='http://localhost:3000'
+
+# Use the development credential lookup strategy
+# ONLY use in development environments
+# Set your public and private keys to "00000000-0000-0000-000000000000" for all requests that should authenticate successfully
+export USE_DEV_KEY_LOOKUP=true
