+++ lib/middleware.js
@@ -9,9 +9,9 @@ module.exports = function( makeModel, apiUserModel, env ) {
       hawkModule = require( "./hawk" )(),
       tags = require( "./tags" )(),
       Make = makeModel,
-      ApiUser = apiUserModel,
       lazyCheck = env.get( "USE_LAZY_ADMIN" ),
-      hawkOptions = {};
+      hawkOptions = {},
+      credentialsLookupStrategy = require( "./strategy" )( apiUserModel, env.get( "USE_DEV_KEY_LOOKUP" ) );
 
   if ( env.get( "FORCE_SSL" ) ) {
     hawkOptions.port = 443;
@@ -80,26 +80,7 @@ module.exports = function( makeModel, apiUserModel, env ) {
       }
     },
     hawkAuth: function( req, res, next ) {
-      hawkModule.Hawk.server.authenticate( req, function ( publickey, callback ) {
-
-        var credentials = {
-          algorithm: 'sha256',
-          user: publickey
-        };
-
-        ApiUser.findOne({ publickey: publickey }, function( err, doc ) {
-          if ( err || !doc ) {
-            return callback( err );
-          }
-          if ( doc.revoked ) {
-            // only warn of revoked key if the passed MAC is successfully authenticated
-            // see the hawk callback below
-            req.revokedKey = true;
-          }
-          credentials.key = doc.privatekey;
-          callback( null, credentials );
-        });
-      }, hawkOptions, function( err, creds, artifacts ) {
+      hawkModule.Hawk.server.authenticate( req, credentialsLookupStrategy, hawkOptions, function( err, creds, artifacts ) {
         var msg;
         if ( err || req.revokedKey ) {
           msg = err ? err.message : "Your Key has been revoked, contact a MakeAPI administrator.";
